scripts/requirements.txt
openai
langchain
langchain-openai
requests
python-dotenv
